+++ src/elements/SendoutElement.php
@@ -48,7 +48,6 @@ use yii\web\Response;
  * @property-read bool $isModifiable
  * @property-read int $segmentCount
  * @property-read SegmentElement[] $segments
- * @property-read int $pendingRecipientCount
  * @property-read bool $isDeletable
  * @property-read int $contactCount
  * @property-read bool $isPausable
@@ -68,6 +67,7 @@ use yii\web\Response;
  * @property-read ContactElement[] $notificationContacts
  * @property-read ContactElement[] $contacts
  * @property-read ContactElement[] $failedContacts
+ * @property-read int $pendingRecipientCount
  * @property-read MailingListElement[] $mailingLists
  */
 class SendoutElement extends Element
@@ -525,11 +525,6 @@ class SendoutElement extends Element
     private ?array $_segments = null;
 
     /**
-     * @var array|null
-     */
-    private ?array $_pendingRecipients = null;
-
-    /**
      * @var ScheduleModel|null Schedule
      * @see getSchedule()
      * @see setSchedule()
@@ -760,7 +755,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the from name, email and reply to.
+     * Returns the “from” name, email and reply to.
      */
     public function getFromNameEmail(): string
     {
@@ -768,7 +763,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the from name, email and reply to label.
+     * Returns the “from” name, email and reply to label.
      */
     public function getFromNameEmailLabel(): string
     {
@@ -782,7 +777,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's progress as a fraction.
+     * Returns the sendout’s progress as a fraction.
      */
     public function getProgressFraction(): float
     {
@@ -790,8 +785,7 @@ class SendoutElement extends Element
             return 1;
         }
 
-        // Get expected recipients
-        $expectedRecipients = $this->getPendingRecipientCount();
+        $expectedRecipients = Campaign::$plugin->sendouts->getPendingRecipientCount($this);
 
         $progress = $expectedRecipients == 0 ?: $this->recipients / ($this->recipients + $expectedRecipients);
 
@@ -799,7 +793,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's progress.
+     * Returns the sendout’s progress.
      */
     public function getProgress(): string
     {
@@ -813,7 +807,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's campaign.
+     * Returns the sendout’s campaign.
      */
     public function getCampaign(): ?CampaignElement
     {
@@ -849,7 +843,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's notification email addresses.
+     * Returns the sendout’s notification email addresses.
      *
      * @return string[]
      */
@@ -866,7 +860,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's notification contacts, or the default notification
+     * Returns the sendout’s notification contacts, or the default notification
      * contacts if this is a fresh sendout.
      *
      * @return ContactElement[]
@@ -885,7 +879,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's contact count.
+     * Returns the sendout’s contact count.
      */
     public function getContactCount(): int
     {
@@ -893,7 +887,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's contacts.
+     * Returns the sendout’s contacts.
      *
      * @return ContactElement[]
      */
@@ -909,7 +903,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's failed contacts.
+     * Returns the sendout’s failed contacts.
      *
      * @return ContactElement[]
      */
@@ -925,7 +919,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's mailing list count.
+     * Returns the sendout’s mailing list count.
      */
     public function getMailingListCount(): int
     {
@@ -933,7 +927,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's mailing lists.
+     * Returns the sendout’s mailing lists.
      *
      * @return MailingListElement[]
      */
@@ -949,7 +943,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's excluded mailing list count.
+     * Returns the sendout’s excluded mailing list count.
      */
     public function getExcludedMailingListCount(): int
     {
@@ -957,7 +951,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's excluded mailing lists.
+     * Returns the sendout’s excluded mailing lists.
      *
      * @return MailingListElement[]
      */
@@ -973,7 +967,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's segment count.
+     * Returns the sendout’s segment count.
      */
     public function getSegmentCount(): int
     {
@@ -981,7 +975,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's segments.
+     * Returns the sendout’s segments.
      *
      * @return SegmentElement[]
      */
@@ -1001,7 +995,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's schedule.
+     * Returns the sendout’s schedule.
      *
      * @return ScheduleModel|null
      */
@@ -1021,30 +1015,28 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's pending recipient contact and mailing list IDs based
+     * Returns the sendout’s pending recipient contact and mailing list IDs based
      * on its mailing lists, segments and schedule.
+     *
+     * @deprecated in 2.13.0. Use `Campaign::$plugin->sendouts->getPendingRecipients()` instead.
      */
-    public function getPendingRecipients(): array
+    public function getPendingRecipients(int $limit = null): array
     {
-        if ($this->_pendingRecipients !== null) {
-            return $this->_pendingRecipients;
-        }
-
-        $this->_pendingRecipients = Campaign::$plugin->sendouts->getPendingRecipients($this);
-
-        return $this->_pendingRecipients;
+        return Campaign::$plugin->sendouts->getPendingRecipients($this, $limit);
     }
 
     /**
-     * Returns the sendout's pending recipient count
+     * Returns the sendout’s pending recipient count.
+     *
+     * @deprecated in 2.13.0. Use `Campaign::$plugin->sendouts->getPendingRecipientCount()` instead.
      */
     public function getPendingRecipientCount(): int
     {
-        return count($this->getPendingRecipients());
+        return Campaign::$plugin->sendouts->getPendingRecipientCount($this);
     }
 
     /**
-     * Returns the sendout's HTML body.
+     * Returns the sendout’s HTML body.
      */
     public function getHtmlBody(): ?string
     {
@@ -1063,7 +1055,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Returns the sendout's plaintext body.
+     * Returns the sendout’s plaintext body.
      */
     public function getPlaintextBody(): ?string
     {
@@ -1171,7 +1163,7 @@ class SendoutElement extends Element
     }
 
     /**
-     * Sets the sendout's schedule.
+     * Sets the sendout’s schedule.
      */
     public function setSchedule(ScheduleModel|array|string|null $schedule): void
     {